fix: tauri window shadows
The app window loses its native borders and shadows when unminimizing it, if window decorations are disabled.

It can be worked around by resizing the window by 1px and immediately shrinking it back when the app unminimizes.

Unfortunately there are no dedicated events for minimize/unminimize available in tauri which we could use to apply this hack, but tauri does fire unrelated events such as `onResized` or `onFocusChanged` when unminimizing.

Tauri does fire absolutely nothing however if the window is minimized while unfocused, eg. when clicking on another app and then clicking on the minimize button.

This commit polls every second whether the app is minimized or not when losing focus - it's not pretty but it works.